Refactor EKS tools: split out deployment, addon, cluster tools
Fixes #5095

- Extracted get_eks_deployment_status, describe_eks_addon, and describe_eks_cluster into dedicated modules
- Extracted shared helpers (_eks_available, _eks_creds) to _shared.py
- Re-exported moved tools in integrations/eks/tools/__init__.py to maintain backwards compatibility
- Added registration smoke test
